Update the README.md file
I have updated the README to have detail on the input file and what the output will be. I also added a How to run section that explains the simple way to run, how to redirect the standard out to a file and how to have it print to screen and file. Lastly, also added a license section.