Add argparse library and set flags for program
Import argparse
Add a Argument Parser with a description.
Add a mutually exclusive group.
Add to the group two optional arguments H(hours) and m(minutes) with the action "store_true".
Add a positional argument file with type FileType('r') default set to '-' which is used for stdin so that when there is no file given to read it will read from stdin instead.
After calling parse_args check the file argument to see if it is reading from stdin or a file
Add logic to check which flag was used, H, m, None
Give the main function the array of lines read and the timeFormat which was decided based on the flags used if any.

Add new function printDayHours which takes a string for the day represented as a letter, a number which is the number of hours for that day, and the timeFormat. The time is calculated as decimal hours and int hours with int minutes. Then it decides which timeFormat is given and prints to screen based on that. The print line for each day in the main function is replaced with a call to this function instead.

Add new function printWeekHours which much like the printDayHours function takes the number of hours and timeFormat. Calculates the hours in different ways the same way. Decides the format the same way and then prints to screen accordingly. The print line for the week is replaced in the main function with a call to this function.