Fix travis coverage report output
The covergae report was giving a report of files other than PunchCard.py which is not what I wanted. So I had to modify .coveragerc to omit tests/*,  */site-packages/* and  */dist-packages/* under the report section and */site-packages/* under the run section. Also changed the comment "Omit tests" to be more specific and read "Omit tests directory".